Refactor message fetching to use updated endpoint and improve recipient options
