+++ t2t-admin/src/services/api.js
@@ -65,18 +65,7 @@ export const apiCall = async (endpoint, method = 'GET', body = null, customHeade
 export const ENDPOINTS = {
     LOGIN: '/login/',
     REGISTER: '/register/',
-    MESSAGES: '/messages/',
+    MESSAGES: '/notifications/',
     SEND_MESSAGE: '/messages/create/',
 };
 
-// Function to fetch messages
-export const fetchMessages = async () => {
-    try {
-        const messages = await apiCall(ENDPOINTS.MESSAGES);
-        console.log('Fetched messages:', messages);
-        return messages;
-    } catch (error) {
-        console.error('Error fetching messages:', error);
-        throw error;
-    }
-};
